Sumo/requirements.txt
absl-py==2.1.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.2
Flask-Cors==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysqlclient==2.2.1
numpy==1.26.4
ortools==9.8.3296
pandas==2.1.3
protobuf==5.26.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1
Werkzeug==3.0.1
xmltodict==0.13.0
traci
paho-mqtt
mariadb
python-socketio[asyncio_client]==5.11.3
sumolib==1.21.0
python-dotenv